+++ src/constants/common.ts
@@ -20,8 +20,13 @@ export const PROTOCOLS = {
   aeternity: 'aeternity',
   solana: 'solana',
   bnb: 'bnb',
+  avalanche: 'avalanche',
+  polygonPos: 'polygonPos',
+  dogecoin: 'dogecoin',
 } as const;
-export const EVM_PROTOCOLS: Protocol[] = [PROTOCOLS.ethereum, PROTOCOLS.bnb];
+export const EVM_PROTOCOLS: Protocol[] = [
+  PROTOCOLS.ethereum, PROTOCOLS.bnb, PROTOCOLS.avalanche, PROTOCOLS.polygonPos,
+];
 
 export const PROTOCOL_LIST = Object.values(PROTOCOLS);